+++ src/parser.ts
@@ -382,8 +382,8 @@ function runifyText(text: string, rainbow = false): Rune[]
 {
 	return replaceAmpCodes(text).replaceAll("\n", "&Tab;\n").replaceAll(/\s*<br>\s*/g, "\n")
 		.replaceAll("—", "-").replaceAll(" - ", " \u2013 ")
-		.replaceAll(/(\s)"(\w)/g, "$1«$2")
-		.replaceAll(/(\w)"(\s)/g, "$1»$2")
+		.replaceAll(/(^|[^\p{L}\d_])"([\p{L}\d_])/gu, "$1«$2")
+		.replaceAll(/([\p{L}\d_])"([^\p{L}\d_]|$)/gu, "$1»$2")
 		.replaceAll(/(^|\s)(\*+)($|\s)/g, sub => sub.replaceAll("*", "&Star;"))
 		.split(/(\[.*\]\(.*\))/g)
 		.map(p =>